HDPI-5180: Move GenAppType to genapp package
