+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/domain/genapp/CitizenGenAppRequest.java
@@ -4,7 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import uk.gov.hmcts.reform.pcs.ccd.domain.GenAppType;
 
 @Builder
 @Data
