+++ src/main/java/uk/gov/hmcts/reform/pcs/ccd/domain/genapp/GenAppType.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.pcs.ccd.domain;
+package uk.gov.hmcts.reform.pcs.ccd.domain.genapp;
 
 public enum GenAppType {
